scripts/requirements.txt
# Python dependencies for migration scripts
anthropic>=0.25.0
python-dotenv>=1.0.0
supabase>=2.0.0
